refactor: match builder filenames to their exported symbol
Swap filenames so each matches its export:
- src/internal/create-retryable.ts            -> createRetryable (deprecated root impl)
- src/<family>-model/create-retryable-model.ts -> createRetryableModel

Tests renamed to follow their subject; import paths updated.
