+++ src/image-model/create-retryable-model.ts
