+++ src/language-model/create-retryable-model.test-d.ts
@@ -7,7 +7,7 @@ import type {
   LanguageModelStream,
   SuccessContext,
 } from '../types.js';
-import { createRetryableModel } from './create-retryable.js';
+import { createRetryableModel } from './create-retryable-model.js';
 
 describe('createRetryableModel', () => {
   it('should return LanguageModel for a model instance', () => {
