+++ src/language-model/index.ts
@@ -6,6 +6,6 @@
  *     from 'ai-retry/language-model';
  */
 
-export { createRetryableModel } from './create-retryable.js';
+export { createRetryableModel } from './create-retryable-model.js';
 
 export * from './conditions/index.js';
